Reorganize salary bot
