--- src/bots/comp/constants/embeds.js
@@ -1,4 +1,0 @@
-import { addSalaryStrings } from './strings';
-import generateEmbed from '../../../utils/embed';
-
-export const addSalaryEmbed = generateEmbed(addSalaryStrings);
