--- src/bots/comp/constants/index.js
@@ -1,34 +1,0 @@
-export const formFields = [
-  {
-    name: 'Company',
-    placeholder: '<company name>'
-  },
-  {
-    name: 'Title',
-    placeholder: '<job title>'
-  },
-  {
-    name: 'Date of Offer',
-    placeholder: '<month year>'
-  },
-  {
-    name: 'Location',
-    placeholder: '<state, city> or <country>'
-  },
-  {
-    name: 'Salary',
-    placeholder: '<salary>'
-  },
-  {
-    name: 'Stock/Recurring Bonus',
-    placeholder: '<stock/annual bonus>'
-  },
-  {
-    name: 'Relocation/Signing Bonus',
-    placeholder: '<one-time bonuses>'
-  },
-  {
-    name: 'Total Compensation',
-    placeholder: '<total yearly compensation>'
-  }
-];
