--- src/bots/comp/constants/strings.js
@@ -1,6 +1,0 @@
-import { description } from './stringHelpers';
-
-export const addSalaryStrings = {
-  description: description(),
-  title: 'Salary Reporter'
-};
