+++ src/bots/comp/embeds.js
@@ -1,0 +1,4 @@
+import { addSalaryStrings } from './constants';
+import generateEmbed from '../../utils/embed';
+
+export const addSalaryEmbed = generateEmbed(addSalaryStrings);
